+++ Tests/SentryTests/Helper/SentryEnabledFeaturesBuilderTests.swift
@@ -25,6 +25,10 @@ final class SentryEnabledFeaturesBuilderTests: XCTestCase {
         options.enablePreWarmedAppStartTracing = true
 #endif // canImport(UIKit)
 #endif // os(iOS) || os(tvOS)
+      
+#if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
+        options.enableAppHangTrackingV2 = true
+#endif //os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
         
         let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
         
@@ -42,5 +46,9 @@ final class SentryEnabledFeaturesBuilderTests: XCTestCase {
         XCTAssert(features.contains("preWarmedAppStartTracing"))
 #endif // canImport(UIKit)
 #endif // os(iOS) || os(tvOS)
+        
+#if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
+        XCTAssert(features.contains("appHangTrackingV2"))
+#endif //os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
     }
 }